+++ Gem/Code/Source/CullingAndLodExampleComponent.cpp
@@ -233,8 +233,8 @@ namespace AtomSampleViewer
             dirLightFP->SetViewFrustumCorrectionEnabled(handle, m_isCascadeCorrectionEnabled);
             dirLightFP->SetShadowFilterMethod(handle, s_shadowFilterMethods[m_shadowFilterMethodIndex]);
             dirLightFP->SetShadowBoundaryWidth(handle, m_boundaryWidth);
-            dirLightFP->SetPredictionSampleCount(handle, m_predictionSampleCount);
-            dirLightFP->SetFilteringSampleCount(handle, m_filteringSampleCount);
+            dirLightFP->SetPredictionSampleCount(handle, static_cast<uint16_t>(m_predictionSampleCount));
+            dirLightFP->SetFilteringSampleCount(handle, static_cast<uint16_t>(m_filteringSampleCount));
             dirLightFP->SetGroundHeight(handle, 0.f);
 
             m_directionalLightHandle = handle;
@@ -363,7 +363,7 @@ namespace AtomSampleViewer
                 ImGui::RadioButton("4", &m_cascadeCount, 4);
             if (cascadesChanged)
             {
-                m_directionalLightFeatureProcessor->SetCascadeCount(m_directionalLightHandle, m_cascadeCount);
+                m_directionalLightFeatureProcessor->SetCascadeCount(m_directionalLightHandle, static_cast<uint16_t>(m_cascadeCount));
             }
 
             ImGui::Spacing();
@@ -407,11 +407,11 @@ namespace AtomSampleViewer
             ImGui::Text("Filtering (PCF specific)");
             if (ImGui::SliderInt("Prediction #", &m_predictionSampleCount, 4, 16))
             {
-                m_directionalLightFeatureProcessor->SetPredictionSampleCount(m_directionalLightHandle, m_predictionSampleCount);
+                m_directionalLightFeatureProcessor->SetPredictionSampleCount(m_directionalLightHandle, static_cast<uint16_t>(m_predictionSampleCount));
             }
             if (ImGui::SliderInt("Filtering #", &m_filteringSampleCount, 4, 64))
             {
-                m_directionalLightFeatureProcessor->SetFilteringSampleCount(m_directionalLightHandle, m_filteringSampleCount);
+                m_directionalLightFeatureProcessor->SetFilteringSampleCount(m_directionalLightHandle, static_cast<uint16_t>(m_filteringSampleCount));
             }
         }
         ImGui::Unindent();
@@ -424,7 +424,7 @@ namespace AtomSampleViewer
             int diskLightCount = m_diskLightCount;
             if (ImGui::SliderInt("Number", &diskLightCount, 0, DiskLightCountMax))
             {
-                UpdateDiskLightCount(diskLightCount);
+                UpdateDiskLightCount(static_cast<uint16_t>(diskLightCount));
             }
 
             if (ImGui::SliderFloat("Intensity##disk", &m_diskLightIntensity, 0.f, 100000.f, "%.1f", ImGuiSliderFlags_Logarithmic))
