+++ Gem/Code/Source/RHI/IndirectRenderingExampleComponent.cpp
@@ -740,7 +740,7 @@ namespace AtomSampleViewer
 
             dispatchItem.m_arguments = dispatchArgs;
             dispatchItem.m_pipelineState = m_cullPipelineState.get();
-            dispatchItem.m_shaderResourceGroupCount = numSrgs;
+            dispatchItem.m_shaderResourceGroupCount = static_cast<uint8_t>(numSrgs);
 
             commandList->Submit(dispatchItem);
         };
@@ -845,7 +845,7 @@ namespace AtomSampleViewer
                 drawItem.m_arguments = m_drawIndirect;
                 drawItem.m_pipelineState = m_drawPipelineState.get();
                 drawItem.m_indexBufferView = &m_indexBufferViews[0];
-                drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                 drawItem.m_shaderResourceGroups = shaderResourceGroups;
                 drawItem.m_streamBufferViewCount = 2;
                 drawItem.m_streamBufferViews = m_streamBufferViews.data();
