+++ Gem/Code/Source/RHI/SubpassExampleComponent.cpp
@@ -412,9 +412,9 @@ namespace AtomSampleViewer
                     drawItem.m_arguments = mesh.m_drawArguments;
                     drawItem.m_pipelineState = modelData.m_pipelineState.get();
                     drawItem.m_indexBufferView = &mesh.m_indexBufferView;
-                    drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+                    drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
                     drawItem.m_shaderResourceGroups = shaderResourceGroups;
-                    drawItem.m_streamBufferViewCount = static_cast<uint32_t>(modelData.m_streamBufferList.size());
+                    drawItem.m_streamBufferViewCount = static_cast<uint8_t>(modelData.m_streamBufferList.size());
                     drawItem.m_streamBufferViews = modelData.m_streamBufferList.data();
 
                     commandList->Submit(drawItem);
@@ -526,7 +526,7 @@ namespace AtomSampleViewer
             drawItem.m_arguments = RHI::DrawArguments(drawArguments);
             drawItem.m_pipelineState = m_compositionPipeline.get();
             drawItem.m_indexBufferView = nullptr;
-            drawItem.m_shaderResourceGroupCount = RHI::ArraySize(shaderResourceGroups);
+            drawItem.m_shaderResourceGroupCount = static_cast<uint8_t>(RHI::ArraySize(shaderResourceGroups));
             drawItem.m_shaderResourceGroups = shaderResourceGroups;
             drawItem.m_streamBufferViewCount = 0;
             drawItem.m_streamBufferViews = nullptr;
